add slf4j and object mapper

+++ src/main/java/com/ria/springweb/service/BirdService.java
@@ -3,7 +3,6 @@ package com.ria.springweb.service;
 import com.ria.springweb.entities.Bird;
 import com.ria.springweb.respository.BirdRespository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.lang.NonNullFields;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -15,33 +14,37 @@ public class BirdService {
 
     @Autowired
     BirdRespository birdRespository;
+    private int id;
 
-    public List<Bird> getBirds() {
+    public List<Bird> GetBirds() {
         return (List<Bird>) birdRespository.findAll();
     }
 
-    public Bird getBird(String name){
+    public Bird GetBird(String name) {
         Bird bird = birdRespository.findByName(name);
-        if(bird==null) throw new NoSuchElementException("Can't find Bird of given name");
+        if (bird == null) throw new NoSuchElementException("Can't find Bird of given name");
         else return bird;
     }
-    public Bird getBird(int id){
-         Optional<Bird> bird = birdRespository.findById(id);
-        if(bird.isPresent()){
+
+    public Bird GetBird(int id) {
+        Optional<Bird> bird = birdRespository.findById(id);
+        if (bird.isPresent()) {
             return bird.get();
-        }
-        else throw new NoSuchElementException("Can't find Bird of given ID");
+        } else throw new NoSuchElementException("Can't find Bird of given ID");
     }
 
-    public Bird addBird(Bird bird) {
+    public Bird AddBird(Bird bird) {
+
         return birdRespository.save(bird);
     }
 
-    public Bird updateBird(Bird bird){
+    public Bird UpdateBird(Bird bird) {
         return birdRespository.save(bird);
     }
 
-    public void deleteBird(int id){
-        birdRespository.deleteById(id);
+    public void DeleteBird(int id) {
+        if (birdRespository.findById(id).isPresent()) birdRespository.deleteById(id);
+        else throw new NoSuchElementException();
+
     }
 }
